refactor package_info
