+++ src/utils/__init__.py
@@ -1,7 +1,3 @@
 from src.utils.create_and_commit_sample_file import create_and_commit_sample_file
-from src.utils.package_info import (
-    extract_info_from_repo_id,
-    package_info_extractor_exceptions,
-)
-from src.utils.retrieve_real_package_name import retrieve_real_package_name
+from src.utils.package_info import extract_info_from_repo_id, retrieve_real_package_name
 from src.utils.standard_log_formatter import StandardLogFormatter
