--- src/utils/package_info/normalize_package_name.py
@@ -1,2 +1,0 @@
-def normalize_package_name(package_name: str) -> str:
-    return package_name.replace("-", "_").replace(".", "_")
