--- src/utils/package_info/package_info_extractor_exceptions.py
@@ -1,10 +1,0 @@
-class PackageInfoExtractorException(Exception):
-    pass
-
-
-class IncorrectURL(PackageInfoExtractorException):
-    pass
-
-
-class InvalidPackageName(PackageInfoExtractorException):
-    pass
